✅ BUILD SUCCESS: Fixed all 76 compilation errors
Integration fixes complete - v1.5.1 now builds successfully!

Error reduction progress:
- Start: 76 errors
- Yesterday: 38 errors (50% reduction)
- This morning: 16 errors
- Final: 0 errors ✅

Fixes applied:
- Added missing System.Linq imports (3 files)
- Added middleware using statements (GetCorrelationId extension)
- Fixed model properties (RewardError: Operation, ErrorMessage, IsResolved, IsRetryable)
- Fixed decimal type conversions in AnalyticsService
- Fixed ErrorStatistics type mapping (Service → ViewModel)
- Fixed EntityFrameworkException → DbUpdateException
- Fixed AutoRecoveryWatchdog GetRetryableErrorsAsync call
- Fixed BitcoinRewardsService LogErrorAsync → LogExceptionAsync
- Fixed ErrorDashboard.cshtml Context rendering
- Added GetSnapshot() method to RewardMetrics

All 6 phases now integrated and building:
- Phase 1: UX improvements (deployed)
- Phase 2: Production hardening
- Phase 3: Testing infrastructure
- Phase 4: Documentation
- Phase 5: Feature parity
- Phase 6: Security audit

Next: Deploy to production

+++ Plugins/BTCPayServer.Plugins.BitcoinRewards/Services/AnalyticsService.cs
@@ -91,7 +91,7 @@ namespace BTCPayServer.Plugins.BitcoinRewards.Services
                     .Where(r => r.Status != RewardStatus.Redeemed)
                     .Sum(r => r.RewardAmountSatoshis),
                 AverageRewardSatoshis = rewards.Count > 0
-                    ? rewards.Average(r => r.RewardAmountSatoshis)
+                    ? (decimal)rewards.Average(r => r.RewardAmountSatoshis)
                     : 0,
                 
                 TotalTransactionAmount = rewards.Sum(r => r.TransactionAmount),
